backend/requirements.txt
altair==6.0.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
attrs==25.4.0
blinker==1.9.0
cachetools==6.2.4
certifi==2026.1.4
charset-normalizer==3.4.4
click==8.3.1
distro==1.9.0
dotenv==0.9.9
fastapi==0.128.0
filelock==3.20.2
fsspec==2025.12.0
gitdb==4.0.12
GitPython==3.1.46
google-auth==2.46.0
google-genai==1.56.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
Jinja2==3.1.6
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
llvmlite==0.46.0
MarkupSafe==3.0.3
more-itertools==10.8.0
mpmath==1.3.0
narwhals==2.15.0
networkx==3.6.1
numba==0.63.1
numpy==2.3.5
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.27.5
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvshmem-cu12==3.3.20
nvidia-nvtx-cu12==12.8.90
openai-whisper==20250625
packaging==25.0
pandas==2.3.3
pillow==12.1.0
protobuf==6.33.2
pyarrow==22.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.12.5
pydantic_core==2.41.5
pydeck==0.9.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-multipart==0.0.21
pytz==2025.2
referencing==0.37.0
regex==2025.11.3
rpds-py==0.30.0
rsa==4.9.1
setuptools==80.9.0
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
starlette==0.50.0
sympy==1.14.0
tenacity==9.1.2
tiktoken==0.12.0
toml==0.10.2
torch==2.9.1
tornado==6.5.4
tqdm==4.67.1
triton==3.5.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.2
uvicorn==0.40.0
watchdog==6.0.0
websockets==15.0.1
